ppt-to-word-ai/requirements.txt
flask==2.3.2
python-pptx==0.6.21
pytesseract==0.3.10
opencv-python-headless==4.7.0.72
python-docx==0.8.11
ollama==0.1.4
pdf2image==1.16.3
Pillow==10.0.0
pymupdf==1.22.5
gunicorn==21.2.0  # Para producción